# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GDColumn是一个基于 Gin 框架搭建的防知乎专栏后端 API。",
        "title": "GDColumn",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "AA12",
            "url": "https://qm.qq.com/cgi-bin/qm/qr?k=qjar90G1gAtMF6uR7-8WLk5OvpD7sjUj\u0026noverify=0",
            "email": "875151567@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login/using-email": {
            "post": {
                "description": "需要用户名密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user 关于用户的路由，登录，注册，获取当前用户等等"
                ],
                "summary": "发送请求用户登录",
                "parameters": [
                    {
                        "description": "用户登录，需要提供用户的邮箱和密码",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.LoginByEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token\":token}",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/auth/signup/using-email": {
            "post": {
                "description": "需要用户名密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user 关于用户的路由，登录，注册，获取当前用户等等"
                ],
                "summary": "注册用户",
                "parameters": [
                    {
                        "description": "用户注册，需要提供用户的邮箱和密码",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.SignupUsingEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            }
        },
        "/columns": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "column 关于专栏的一些操作信息"
                ],
                "summary": "发送请求，获得专栏列表",
                "parameters": [
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/column.Column"
                        }
                    }
                }
            }
        },
        "/columns/": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "column 关于专栏的一些操作信息"
                ],
                "summary": "发送请求，更新个人专栏信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "待更新专栏数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.ColumnRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/column.Column"
                        }
                    }
                }
            }
        },
        "/columns/{id}": {
            "get": {
                "tags": [
                    "column 关于专栏的一些操作信息"
                ],
                "summary": "发送请求，获取一个专栏详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "专栏ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/column.Column"
                        }
                    }
                }
            }
        },
        "/columns/{id}/posts": {
            "get": {
                "description": "专栏ID",
                "tags": [
                    "post 关于文章的一些操作信息"
                ],
                "summary": "发送请求获取属于专栏的文章信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "专栏ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/post.Post"
                        }
                    }
                }
            }
        },
        "/current": {
            "get": {
                "tags": [
                    "user 关于用户的路由，登录，注册，获取当前用户等等"
                ],
                "summary": "获取当前用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            }
        },
        "/posts": {
            "post": {
                "description": "需要标题、内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post 关于文章的一些操作信息"
                ],
                "summary": "新建文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "新建文章，需要提供标题和内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.PostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/post.Post"
                        }
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "description": "文章ID",
                "tags": [
                    "post 关于文章的一些操作信息"
                ],
                "summary": "发送请求获取单个文章信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/post.Post"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post 关于文章的一些操作信息"
                ],
                "summary": "发送请求，更新单个文章信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "待更新文章数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.PostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/post.Post"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post 关于文章的一些操作信息"
                ],
                "summary": "发送请求，删除单个文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/post.Post"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "专栏ID",
                "tags": [
                    "file 关于文件上传的一些操作信息"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/image.Image"
                        }
                    }
                }
            }
        },
        "/users": {
            "put": {
                "tags": [
                    "user 关于用户的路由，登录，注册，获取当前用户等等"
                ],
                "summary": "更新当前用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "待更新用户数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserUpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "column.Column": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "avatar": {
                    "$ref": "#/definitions/column.Image"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "column.Image": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "image.Image": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "post.Image": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "post.Post": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/user.User"
                },
                "columnId": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "$ref": "#/definitions/post.Image"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "requests.ColumnRequest": {
            "type": "object",
            "properties": {
                "avatarId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "requests.LoginByEmailRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "requests.PostRequest": {
            "type": "object",
            "properties": {
                "authorId": {
                    "type": "string"
                },
                "columnId": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "requests.SignupUsingEmailRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "password_confirm": {
                    "type": "string"
                }
            }
        },
        "requests.UserUpdateProfileRequest": {
            "type": "object",
            "properties": {
                "avatarId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                }
            }
        },
        "user.Image": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "user.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "$ref": "#/definitions/user.Image"
                },
                "columnId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}